Use common warnings in more places
Create a CMake function to apply warnings more consistently. Apply to:

- capture_service
- gfxr_ext/decode
- gpu_time
- ui

Apply at using directory-based commands so that we can't forget to apply
the config to a particular target.

This should catch more bugs.

+++ capture_service/android_application.cc
@@ -85,7 +85,7 @@ namespace Dive
 int GetIndentation(const std::string& line)
 {
     int indention = 0;
-    for (int i = 0; i < line.size(); i++)
+    for (int i = 0; i < static_cast<int>(line.size()); i++)
     {
         if (line[i] == ' ')
         {
